♻️ Refactor: 챗봇 타입 분리
